Only disconnect when connection is down

+++ roles/workstation.nix
@@ -84,7 +84,7 @@ in {
                 logger "BK-G UP: Adding static route to amon $AMON_IP via $IP4_GATEWAY"
                 ip r a $AMON_IP via $IP4_GATEWAY || true
                 systemctl start openvpn-tcp-to-lanfear.service || true
-              else
+              elif [[ "$NM_DISPATCHER_ACTION" == "down" ]]; then
                 logger "BK-G DOWN: Deleting static route to amon $AMON_IP"
                 ip r d $AMON_IP via $IP4_GATEWAY || true
                 systemctl stop openvpn-tcp-to-lanfear.service || true
